+++ lib/chatterbox/services/email.rb
@@ -4,20 +4,38 @@ require 'chatterbox/services/email/mailer'
 module Chatterbox::Services
   class Email
     attr_reader :options
-    @default_configuration = {}
-
+    # Return default configuration - defaults to empty hash
     def self.default_configuration
-      @default_configuration
+      @default_configuration ||= {}
     end
-    
+
+    # Configure Email service with default options
     def self.configure(config_options)
-      @default_configuration = config_options
+      @default_configuration = config_options.with_indifferent_access
     end
     
+    # Deliver the a notification via email
+    #
+    # ==== Options
+    #
+    # * :summary - The subject of the message - required.
+    # * :body - The body of the email - if none provided, you will just send an email with a blank body.
+    # * :config - A sub-hash of configuration options.  
+    #   Optional if you have configured a default configuration with #configure.
+    #   Config can contain any of the following options, all of which are pretty self explanatory:
+    #
+    #     :to
+    #     :from
+    #     :content_type
+    #     :reply_to
+    #     :bcc
+    #     :cc
     def self.deliver(options = {})
       new(options).deliver
     end
     
+    # Creates a new Email service with provided options.
+    # See Email::deliver for the valid options
     def initialize(options = {})
       @options = options
       merge_configs
@@ -25,6 +43,7 @@ module Chatterbox::Services
       validate_options
     end
     
+    # Deliver a notificaiton -- normally you should just use the class level method Email::deliver
     def deliver
       Mailer.deliver_message(options)
     end